Added search functionality.
Uses nested for loops that go through the ListViewItem's sub-items, and then tag the item as a match if the query string matches the sub-item text content value. If an item doesn't have a match tag, then it is removed, effectively filtering out items that don't match, and leaving behind those that do.