Property ListView is now populated, and a whole lot more.
District popup no longer lets the user specify the number of neighborhoods as this will be done automatically.

Neighborhood popup now includes the district name so the user can specify which district they want to add the neighborhood to. Selecting another neighborhood first before clicking on add autofills the district name.

Property popup now features both the district name and the neighborhood name so the user knows which neighborhood of which district they are adding a property to.