Changed how the data is stored after being read.
Initially, I was using an array of strings for each line of the data file. After looking at the "Books" and "Library" applications, an overaching "District[]" array is now used to store the data. This is a lot more object oriented, and it's way easier to modify. It's also easier to loop through to display the data. Plus, I think this is how we were meant to do it anyway.