Added three methods: DisableControl, EnableControl, and AddColumns.
DisableControl and EnableControl are both part of the Utils class, and they simply change the background and foreground colors of controls to make them appear disabled or enabled. I didn't like the appearance of a button that is actually disabled, and changing that is a lot of hassle, so I went with this approach.

The AddColumns method takes a category as a parameter, and then populates a ListView based on the category with columns.