ListViews are no longer laggy to interact with as the ".Clear()" method is now used.
Looked at the "Books" and "Libraries" applications and found out that the "Clear()" method is used and isn't considered a shortcut. It's much faster, so that is now used to clear ListViews. ListViews now only populate if a parent item is selected (click on a district to view its neighborhoods, click on a neighborhood to view its properties). Showing all neighborhoods and properties as a sort of overview caused a lot of lag.