Static ListView item indexing, validation checks, and more search functionality.
The assignment specifications first state that neighborhoods and properties need to be able to be selected, but not searched, but then at the end says they need to be able to be searched for. As a result, I added a search functionality for all three categories just in case.

ListView items now have an extra (but hidden) column that contains the item's index. This is now used instead of the "SelectedIndices" class member of the ListView class. This way, when the user searches for something and the other items are removed to isolate the desired item, the item retains its original index.

Validation checks have been added to the property popup form.